pyproject.toml
[project]
name = "django-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "robertmertens10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "environs (>=14.1.1,<15.0.0)"
]

[tool.poetry]
packages = [{include = "django_test", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipython = "^9.2.0"
black = "^25.1.0"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
